services/chatbot-service/requirements.txt

fastapi
uvicorn
python-dotenv
google-generativeai
sqlalchemy
psycopg2-binary
